finish navbar
