+++ pages/_app.tsx
@@ -8,7 +8,6 @@ import "src/components/BoardGrid/styles.scss";
 import "src/components/Button/styles.scss";
 import "src/components/CommunitiesGrid/styles.scss";
 import "src/components/Footer/styles.scss";
-import "src/components/NavigationBar/styles.scss";
 import "src/components/ScrollDownArrow/styles.scss";
 import "src/components/Statistic/styles.scss";
 // section css imports
@@ -17,7 +16,7 @@ import "src/sections/community/styles.scss";
 import "src/sections/home/styles.scss";
 import "src/sections/sponsorship/styles.scss";
 
-import NavigationBar from "src/components/NavigationBar/nav2";
+import NavigationBar from "src/components/NavigationBar";
 import Footer from "src/components/Footer";
 
 const MyApp = ({ Component, pageProps }) => (
