+++ src/components/NavigationBar/index.tsx
@@ -1,89 +1,75 @@
 import Link from "next/link";
-import Image from "next/image";
-import { useState } from "react";
+import { useEffect, useState } from "react";
 import ACMLogo from "public/assets/ACMWhiteLogo.png";
-import CloseMenuButton from "public/assets/closemenubutton.svg";
-import MenuButtonIcon from "public/assets/menubutton.svg";
+import s from "src/styles/modules/App.module.scss";
+import { Size, useWindowSize } from "src/utils";
 
 const navLinks = [
+  { to: "#contact", text: "Contact" },
   { to: "/sponsor", text: "Sponsor" },
   { to: "/communities", text: "Communities" },
   { to: "/about", text: "About Us" },
 ];
 
 const NavigationBar: React.FC = () => {
-  const [menuState, setMenuState] = useState(false);
+  const size: Size = useWindowSize();
 
-  const toggleMenu = () => setMenuState(!menuState);
+  const [menuOpen, setMenuOpen] = useState(false);
+  const [mobile, setMobile] = useState(false);
+  const toggleMenu = () => setMenuOpen(!menuOpen);
+
+  // Switch to mobile less than 920px
+  useEffect(() => {
+    setMobile(size.width < 920);
+  }, [size]);
+
+  // If they go back to desktop size, don't keep the menu open
+  useEffect(() => {
+    if (!mobile) setMenuOpen(false);
+  }, [mobile]);
 
   return (
-    <div className="fixed-nav">
-      <div className="navigation-bar">
-        <ul>
-          <li className="navigation-bar__logo">
-            <Link href="/">
-              <a>
-                <img src={ACMLogo.src} alt="ACM Logo" />
-                <p>at UCSD</p>
-              </a>
-            </Link>
-          </li>
-          <li className="navigation-bar__login">
-            <Link href={"https://members.acmucsd.com/login"}>
-              <a>Member Login</a>
-            </Link>
-          </li>
-          <li className="navigation-bar__nav-button">
-            <Link href="#contact">
-              <a>Contact</a>
-            </Link>
-          </li>
+    <div className={s.navbarWrapper}>
+      <div className={s.navbar}>
+        {/* Navbar ACM Logo */}
+        <div className={s.left}>
+          <Link href="/">
+            <a>
+              <img src={ACMLogo.src} alt="ACM Logo" />
+              <p>at UCSD</p>
+            </a>
+          </Link>
+        </div>
+
+        {/* Desktop Nav Links */}
+        <div className={`${s.right} ${mobile && s.hidden}`}>
           {navLinks.map((link, key) => (
-            <li key={key} className="navigation-bar__nav-button">
-              <Link href={link.to}>
-                <a>{link.text}</a>
-              </Link>
-            </li>
+            <Link key={key} href={link.to}>
+              <a className={s.navItem}>{link.text}</a>
+            </Link>
           ))}
-        </ul>
+          <Link href={"https://members.acmucsd.com/login"}>
+            <a className={s.loginButton}>Member Login</a>
+          </Link>
+        </div>
+        {/* Mobile Navbar Toggle */}
+        <button className={`${s.toggleIcon} ${!mobile && s.hidden}`} onClick={toggleMenu}>
+          <div className={`${s.bar1} ${menuOpen && s.open}`} />
+          <div className={`${s.bar2} ${menuOpen && s.open}`} />
+        </button>
       </div>
 
-      {menuState ? (
-        <div className="mobile-links">
-          <button className="closebutton" onClick={toggleMenu} type="button">
-            <img className="closeicon" src={CloseMenuButton.src} color="white" alt="Menu Icon" />
-          </button>
-          <Link href="/about">
-            <a onClick={toggleMenu}>About Us</a>
+      {/* Mobile Menu Dropdown */}
+      <div className={`${s.mobileNav} ${menuOpen && s.open}`}>
+        {navLinks.map((link, key) => (
+          <Link key={key} href={link.to}>
+            <a className={`${s.navItem}`}>{link.text}</a>
           </Link>
-          <Link href="/communities">
-            <a onClick={toggleMenu}>Communities</a>
-          </Link>
-          <Link href="/sponsor">
-            <a onClick={toggleMenu}>Sponsor</a>
-          </Link>
-          <Link href="#contact">
-            <a onClick={toggleMenu}>Contact</a>
-          </Link>
-          <a href="https://members.acmucsd.com/login">Login</a>
-        </div>
-      ) : (
-        <div className="navbar-mobile-div">
-          <div className="navbar-mobile">
-            <Link href="/">
-              <a>
-                <img src={ACMLogo.src} alt="ACM Logo" />
-                <p>at UCSD</p>
-              </a>
-            </Link>
-            <button className="icon" onClick={toggleMenu} type="button">
-              <img className="menuicon" src={MenuButtonIcon.src} alt="Menu Icon" />
-            </button>
-          </div>
-        </div>
-      )}
+        ))}
+      </div>
 
-      <div className="rainbow" />
+      {/* Bottom Rainbow */}
+      <div className={s.rainbow} />
     </div>
   );
 };
